optimize

+++ deploy.sh
@@ -17,31 +17,27 @@ USER=centos
 PASSWORD=123456
 codePath=/home/work/code/s90_platform
 
-# start SSH-Agent and get its pid
-sshAgentPid=$(eval `ssh-agent` | grep -Eo "[0-9]+")
-sshAddResult=$(ssh-add ~/.ssh/phab)
-echo "ssh agent pid: ${sshAgentPid}"
-echo "ssh file added result: ${sshAddResult}"
-
-if [[ $1 =~ [a] ]]
-then
- echo "ssh agent start authenticated file added"
- exit 0
+if [[ $1 =~ [a] ]]; then
+    # start SSH-Agent and get its pid
+    sshAgentPid=$(eval $(ssh-agent) | grep -Eo "[0-9]+")
+    sshAddResult=$(ssh-add ~/.ssh/phab)
+    echo "ssh agent pid: ${sshAgentPid}"
+    echo "ssh file added result: ${sshAddResult}"
+    echo "ssh agent start authenticated file added"
+#  exit 0
 fi
 
 cd ${codePath} || echo 'cd code path failed' || exit 1
 pwd
 
-if [[ $1 =~ [i] ]]
-then
- echo "run install"
- npm install
+if [[ $1 =~ [i] ]]; then
+    echo "run install"
+    npm install
 fi
 
-if [[ $1 =~ [p] ]]
-then
- echo "run pack"
- npm run pack:linux
+if [[ $1 =~ [p] ]]; then
+    echo "run pack"
+    npm run pack:linux
 fi
 
 version=$(cat <${codePath}/package.json | jq '.version')
@@ -67,6 +63,10 @@ tar -xzvf ${filename}
 exit
 EOF
 
-kill -9 ${sshAgentPid}
+if [[ $1 =~ [a] ]]; then
+    # kill SSH-Agent
+    kill -9 ${sshAgentPid}
+    echo "ssh agent pid: ${sshAgentPid} killed"
+fi
 
 echo "update ${TARGETIP} ${filename} finished"
